chore: move sample file
